Improve build caching (#599)
* Support caching containerd builds

If we specify the path here, then snapcraft doesn't think that the
containerd part is built from the microk8s tree any more, but rather is
dependent on the source code that's pulled from github. This allows
snapcraft part build caching to work correctly for this part.

Fixes: #586

Signed-off-by: Joe Stringer <joe@cilium.io>

* Support populating KUBE_SNAP_BINS from scratch

This commit improves the development experience by allowing a developer
to simply export KUBE_SNAP_BINS=/some/path then if the path doesn't
exist, the build will populate that path with the k8s artifacts and use
them. If the path already exists, it will use the artifacts there. The
latter case was already supported, this commit supports the former case.

Previously, if KUBE_SNAP_BINS was specified and the path didn't already
have the cached artifacts, the build would fail; to actually cache the
artifacts (and not delete them), the developer would have to manually
edit the snapcraft.yaml to ensure that the artifacts aren't cleaned up,
and then start using them.

This greatly speeds up the creation of new microk8s snaps in my
environment.